refactor: Change option prefixes to end on a whitespace
This change fixes bug where parser previously cannot distinguish
between `-p` and `-pp`. Now, the parser can differentiate between
`-p ` and `-pp `, as `-p ` is not a front substring of `-pp ` due to
the whitespace char at the end of `-p `.

Accordingly, AddressBookParser has been modified such that it only
removes leading whitespaces, and no longer removes trailing
whitespaces. This is to account for optional options such as `-r `,
where `-r ` will not be registered as a valid key since the
whitespace char is trimmed away before the parsing of keys.

Not removing trailing whitespaces should not affect the eventual
values of the keys, as tokenizer will still trim the parsed value
before returning it. ie `-a test street     ` will still result in
the address value being `test street`, without trailing spaces.